Structural file updates
