+++ cmd/iconbridge/chain/algo/sender.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"math/big"
 	"time"
 
@@ -88,50 +87,7 @@ type relayTx struct {
 	cl     IClient
 }
 
-func (s *sender) initAbi() error {
-	rawBmc, err := ioutil.ReadFile(contractDir)
-	if err != nil {
-		return fmt.Errorf("Failed to open contract file: %w", err)
-	}
-	abiBmc := &abi.Contract{}
-	if err = json.Unmarshal(rawBmc, abiBmc); err != nil {
-		return fmt.Errorf("Failed to marshal abi contract: %w", err)
-	}
-	s.bmc = abiBmc
-	sp, err := s.cl.algod.SuggestedParams().Do(context.Background())
-	if err != nil {
-		return fmt.Errorf("Failed to get suggeted params: %w", err)
-	}
-	sp.Fee = 0
-	s.mcp = &future.AddMethodCallParams{
-		AppID:           s.opts.AppId,
-		Sender:          s.wallet.TypedAddress(),
-		SuggestedParams: sp,
-		OnComplete:      types.NoOpOC,
-		Signer:          s.wallet,
-	}
-	return nil
-}
-
-func (s *sender) CallAbi(name string, args []interface{}) (future.ExecuteResult, error) {
-	var atc = future.AtomicTransactionComposer{}
-	method, err := getMethod(s.bmc, name)
-	if err != nil {
-		return future.ExecuteResult{}, fmt.Errorf("Failed to get %s method from json contract: %w",
-			name, err)
-	}
-	err = atc.AddMethodCall(combine(*s.mcp, method, args))
-	if err != nil {
-		return future.ExecuteResult{}, fmt.Errorf("Failed to add %s method to atc: %w", name, err)
-	}
-	ret, err := atc.Execute(s.cl.algod, context.Background(), 5)
-	if err != nil {
-		return future.ExecuteResult{}, fmt.Errorf("Failed to execute atc: %w", err)
-	}
-	return ret, nil
-}
-
-func (opts *senderOptions) Unmarshal(v map[string]interface{}) error {
+func (opts *senderOptions) unmarshal(v map[string]interface{}) error {
 	b, err := json.Marshal(v)
 	if err != nil {
 		return err
